requirements.txt
Pillow
simpleaudio
numpy
RPi.GPIO
spidev